\name{classifyType}
\alias{classifyType}
\title{classifyType}
\description{Classifies attributes according to R data types}
\usage{classifyType(elBaseType)}
\arguments{
  \item{elBaseType}{The xs: namespace-specific definition name of the value}
}



\author{Lucy Bastin and Ankita Goyal}





