\name{getBaseType}
\alias{getBaseType}
\title{getBaseType}
\description{Gives the name of the base type for the given definition of the data type}
\usage{getBaseType(theDoc, typeDef, namespaces)}
\arguments{
  \item{theDoc}{All or part of the XML Schema document containing the given type definition}
  \item{typeDef}{The given type definition}
  \item{namespaces}{The namespaces present in the document}
}

\value{Returns the name of the base type of the element}

\author{Lucy Bastin and Ankita Goyal}





