\name{getElementDetails}
\alias{getElementDetails}
\title{getElementDetails}
\description{Gives the details of the given element}
\usage{getElementDetails(theEl, theDoc, namespaces, logLevel = 2)}
\arguments{
  \item{theEl}{The element whose details are needed}
  \item{theDoc}{All or part of the XML Schema document containing the element and its definition}
  \item{namespaces}{The namespaces present in the document}
  \item{logLevel}{The level of logging that will be carried out: 0 (none) 1 (limited) or 2 (full). Optional - defaults to 0}
}

\value{Returns a list giving all the details of the given element}

\author{Lucy Bastin and Ankita Goyal}





