\name{getElementExtensionBase}
\alias{getElementExtensionBase}
\title{getElementExtensionBase}
\description{Gives the base type of the extension in an element}
\usage{getElementExtensionBase(element, namespaces, logLevel = 2)}
\arguments{
  \item{element}{The element for which the base type of the extension is to be found}
  \item{namespaces}{The namespaces present in the document containing the element}
  \item{logLevel}{The level of logging that will be carried out: 0 (none) 1 (limited) or 2 (full). Optional - defaults to 0}
}

\value{Returns the base type of the extension in an element, if an extension is present, NULL otherwise}

\author{Lucy Bastin and Ankita Goyal}





