\name{getElementListBase}
\alias{getElementListBase}
\title{getElementListBase}
\description{Gives the data type of the items in the element's list }
\usage{getElementListBase(element, namespaces, logLevel = 2)}
\arguments{
  \item{element}{The element which contains a list}
  \item{namespaces}{The namespaces present in the document containing the element}
  \item{logLevel}{The level of logging that will be carried out: 0 (none) 1 (limited) or 2 (full). Optional - defaults to 0}
}

\value{Returns the data type of the items that the list contains, if the given element contains a list, NULL otherwise}

\author{Lucy Bastin and Ankita Goyal}





