\name{getMinMaxValues}
\alias{getMinMaxValues}
\title{getMinMaxValues}
\description{Collects details of min/max value restrictions on a supplied element}
\usage{getMinMaxValues(inDef, namespaces)}
\arguments{
  \item{inDef}{The element to be queried}
  \item{namespaces}{The relevant namespaces for the document}
}



\author{Lucy Bastin and Ankita Goyal}





