\name{getNamespaceDefinitions}
\alias{getNamespaceDefinitions}
\title{getNamespaceDefinitions}
\description{Gives the namspace definitions present in the given XML Schema document}
\usage{getNamespaceDefinitions(schemaDoc)}
\arguments{
  \item{schemaDoc}{The name of the XML Schema document}
}

\value{Returns the namespace definitions in the given XML Schema document}

\author{Lucy Bastin and Ankita Goyal}





