\name{getNamespaces}
\alias{getNamespaces}
\title{getNamespaces}
\description{Gives the namspace definitions present in the given XML Schema document}
\usage{getNamespaces(doc)}
\arguments{
  \item{doc}{The name of the XML Schema document}
}

\value{Returns the namespace definitions in the given XML Schema document}

\author{Lucy Bastin and Ankita Goyal}





