\name{getTypeDefinitionFromTypeDefName}
\alias{getTypeDefinitionFromTypeDefName}
\title{getTypeDefinitionFromTypeDefName}
\description{Gives the definition for the given data type}
\usage{getTypeDefinitionFromTypeDefName(typeDefName, theDoc, namespaces, 
    logLevel = 2)}
\arguments{
  \item{typeDefName}{The name of the data type}
  \item{theDoc}{The XML Schema document containing the data type and its definition}
  \item{namespaces}{The namespaces present in the document}
  \item{logLevel}{The level of logging that will be carried out: 0 (none) 1 (limited) or 2 (full). Optional - defaults to 0}
}

\value{Return the type definition of the given data type}

\author{Lucy Bastin and Ankita Goyal}





