\name{schemaReader-package}
\alias{schemaReader-package}
\docType{package}
\title{Tools for reading XML Schema within R}
\description{Reads simple XML schema and returns details of user-specified elements}
\details{
\tabular{ll}{Package: \tab schemaReader\cr
Version: \tab 0.1.0\cr
Date: \tab 2012-06-27\cr
Title: \tab Tools for reading XML Schema within R\cr
Author: \tab Lucy Bastin and Ankita Goyal\cr
Maintainer: \tab Lucy Bastin and Ankita Goyal\cr
Depends: \tab R (>= 2.5.0), XML\cr
License: \tab GPL-2\cr
Packaged: \tab 2012-06-27 17:34:06\cr
Repository: \tab CRAN\cr
Date/Publication: \tab 2012-06-27 17:34:06\cr
Built: \tab R 2.9.2; ; 2012-06-27 17:34:06 UTC; windows\cr}
}
\author{Lucy Bastin and Ankita Goyal}

\keyword{ package }


