\name{getDirectTypeDefForElement}
\alias{getDirectTypeDefForElement}
\title{getDirectTypeDefForElement}
\description{Gives the definition of the data type of the element, given the element}
\usage{getDirectTypeDefForElement(en, theDoc, namespaces)}
\arguments{
  \item{en}{The element in the XML Schema document, whose type definition is required}
  \item{theDoc}{The XML Schema document containing the element}
  \item{namespaces}{The namespaces present in the document}
}

\value{Returns the definition of the data type of the given element in the given XML Schema document}

\author{Lucy Bastin and Ankita Goyal}





