\name{getDirectTypeNameForElement}
\alias{getDirectTypeNameForElement}
\title{getDirectTypeNameForElement}
\description{Gives the name of the data type of the given element}
\usage{getDirectTypeNameForElement(en, theDoc, namespaces, logLevel = 2)}
\arguments{
  \item{en}{The element, the name of whose data type is required}
  \item{theDoc}{The XML Schema document containing the element}
  \item{namespaces}{The namespaces present in the document}
  \item{logLevel}{The level of logging that will be carried out: 0 (none) 1 (limited) or 2 (full). Optional - defaults to 0}
}

\value{Returns the name of the data type of the element}

\author{Lucy Bastin and Ankita Goyal}





