\name{getElementDetails}
\alias{getElementDetails}
\title{getElementDetails}
\description{Gives the details of the given element}
\usage{getElementDetails(theEl, theDoc, namespaces)}
\arguments{
  \item{theEl}{The element whose details are needed}
  \item{theDoc}{All or part of the XML Schema document containing the element and its definition}
  \item{namespaces}{The namespaces present in the document}
}

\value{Returns a list giving all the details of the given element}

\author{Lucy Bastin and Ankita Goyal}





