\name{getElementExtensionBase}
\alias{getElementExtensionBase}
\title{getElementExtensionBase}
\description{Gives the base type of the extension in an element}
\usage{getElementExtensionBase(element, namespaces)}
\arguments{
  \item{element}{The element for which the base type of the extension is to be found}
  \item{namespaces}{The namespaces present in the document containing the element}
}

\value{Returns the base type of the extension in an element, if an extension is present, NULL otherwise}

\author{Lucy Bastin and Ankita Goyal}





