\name{getElementListBase}
\alias{getElementListBase}
\title{getElementListBase}
\description{Gives the data type of the items in the element's list }
\usage{getElementListBase(element, namespaces)}
\arguments{
  \item{element}{The element which contains a list}
  \item{namespaces}{The namespaces present in the document containing the element}
}

\value{Returns the data type of the items that the list contains, if the given element contains a list, NULL otherwise}

\author{Lucy Bastin and Ankita Goyal}





